Add bip32 keys and test vectors
